refactor : 상품디테일 페이지 컴포넌트로 분리
